QPID-5612: JMS benchmarking tool [Part of the existing Java Broker Performance Test suite]
Simple tool that uses the existing performance test suite to give a message throughput statistics (msg/s) and (Kbytes/s)
for a use case involving a producer/consumer (on separate connections) with persistent messages on transactional
sessions.  The test scales the number of connections through 1, 2, 5 and 10 and reports separate statistics for
each.  The duration of the test and message size can be overridden from the command line via system properties
-Dqpid.disttest.duration and -Dqpid.disttest.messageSize respectively.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@1575336 13f79535-47bb-0310-9956-ffa450edef68