QPID-6350: [Java Common] Change IoSender to half close  once writing is done.
Change IoSender half close once it knows it has finished writing data down the pipe.  This is done
to prevent the peer from seeing sporadic socket exception 'connection resets' if it happens that the other
side closes the socket first.  If half close is not supported (windows platform or SSLSocket), this change
has no effect.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@1657313 13f79535-47bb-0310-9956-ffa450edef68