QPID-6775: [Java Broker] Avoid immediate reload of metadata after flow to disk.
Be flowing to disk after enqueuing we avoid the reload of the metadata that
occurs on calling isPersistent() in the queueEntry creation.
This is an improvement but might not fully address the problem but the memory
impact of putting a persistent flag into ServerMessages is considered to
intrusive at this point in the release cycle.

git-svn-id: https://svn.apache.org/repos/asf/qpid/java/trunk@1709987 13f79535-47bb-0310-9956-ffa450edef68