QPID-6997: [Java Broker] BDB JE HA: Ensure that the VH is closed before closing the replicated environment
Ensure that the VH is closed *before* starting to restart the environment. Closing the virtualhost will not
return until connections are closed, and closing connections is guaranteed to rollback in flight transactions.
This will ensure that we comply with JE's business rule that transactions/cursors must be closed before
closing the environment.

git-svn-id: https://svn.apache.org/repos/asf/qpid/java/trunk@1726358 13f79535-47bb-0310-9956-ffa450edef68